backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
hypothesis==6.122.3
pytesseract==0.3.13
Pillow>=10.4.0
python-multipart==0.0.20
